Add support for Oscar 0.7 and 0.8
Two major changes in Oscar's version 0.6 and 0.8 respectively, require
changes to the calculation of total costs of a basket. To support this
without cluttering the code base with more and more version-related
checks. This commit adds a compat layer for the payment detail view that
allows better handling of various Oscar versions.